Add an if statement around executing npm

+++ npm/recipes/default.rb
@@ -1,7 +1,9 @@
 include_recipe 'deploy'
 
-node[:deploy].each do |application, deploy|
-  execute "/usr/local/bin/npm install --production" do
-    cwd "#{deploy[:deploy_to]}/current"
+if File.exists?("#{deploy[:deploy_to]}/current/package.json")
+  node[:deploy].each do |application, deploy|
+    execute "/usr/local/bin/npm install --production" do
+      cwd "#{deploy[:deploy_to]}/current"
+    end
   end
 end
